ai_engine/requirements.txt
requests>=2.25.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
web3>=6.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
